Updates in gMSM scripts.

+++ gMSM_scripts/cross_reg_dataset.sh
@@ -5,7 +5,7 @@
 #SBATCH --nodes=1
 #SBATCH --mem=1024
 #SBATCH --ntasks=1
-#SBATCH --job-name=cgMSM
+#SBATCH --job-name=cgMSMctr
 
 dataset=HCP
 workdir=/scratch/users/k2258483/groupwise/${dataset}
@@ -18,15 +18,27 @@ do
 	
 	if [ $(head -c 1 $blocks_folder/block_$filenum.txt) -eq 0 ]
 	then
-	    echo "sbatch --wait --time=0-1:00 --ntasks=4 --array=1-${filelen} --output=${workdir}/logs/groupwise_bunch_${filenum}_%a.txt --export=ALL,FILENUM=${filenum} cross_register.sh"
-		time sbatch --wait --time=0-1:00 --ntasks=4 --array=1-${filelen} --output=${workdir}/logs/groupwise_bunch_${filenum}_%a.txt --export=ALL,FILENUM=${filenum} cross_register.sh
+	    echo "sbatch --wait --time=0-1:00 --ntasks=2 --array=1-${filelen} --output=${workdir}/logs/groupwise_bunch_${filenum}_%a.txt --export=ALL,FILENUM=${filenum} cross_register.sh"
+		time sbatch --wait --time=0-1:00 --ntasks=2 --array=1-${filelen} --output=${workdir}/logs/groupwise_bunch_${filenum}_%a.txt --export=ALL,FILENUM=${filenum} cross_register.sh
 	fi
+
+    retval=$?
+    if [ $retval -ne 0 ]; then
+        echo "Error at step $filenum"
+        exit $retval
+    fi
 
 	if [ $(head -c 1 $blocks_folder/block_$filenum.txt) -eq 1 ]
 	then
-	    echo "sbatch --wait --time=0-3:00 --ntasks=1 --array=1-${filelen} --output=${workdir}/logs/groupwise_bunch_${filenum}_%a.txt --export=ALL,FILENUM=${filenum} cross_register.sh"
-		time sbatch --wait --time=0-3:00 --ntasks=1 --array=1-${filelen} --output=${workdir}/logs/groupwise_bunch_${filenum}_%a.txt --export=ALL,FILENUM=${filenum} cross_register.sh
+	    echo "sbatch --wait --time=0-1:00 --ntasks=1 --array=1-${filelen} --output=${workdir}/logs/groupwise_bunch_${filenum}_%a.txt --export=ALL,FILENUM=${filenum} cross_register.sh"
+		time sbatch --wait --time=0-1:00 --ntasks=1 --array=1-${filelen} --output=${workdir}/logs/groupwise_bunch_${filenum}_%a.txt --export=ALL,FILENUM=${filenum} cross_register.sh
 		cat $workdir/new_clusters/frontal_subject_clusters_HCP_* >> $workdir/frontal_subject_clusters_HCP.csv
 		rm $workdir/new_clusters/frontal_subject_clusters_HCP_*
 	fi
+
+    retval=$?
+    if [ $retval -ne 0 ]; then
+        echo "Error at step $filenum"
+        exit $retval
+    fi
 done
